Add Recover Pass and Premium

+++ src/repositories/product.repository.js
@@ -34,7 +34,7 @@ class ProductRepository{
         
     } 
     
-    async addProduct(title,description,price,code,stock, category, status = true, thumbnails = []){
+    async addProduct(title,description,price,code,stock, category, status = true, thumbnails = [], ownerId = "admin"){
         /* if(!title || !description || !price || !code || !stock || !category){
         CustomError.createError({
             name: "Product creation error",
@@ -43,9 +43,10 @@ class ProductRepository{
             code: EnumErrors.INVALID_TYPES_ERROR
         })
     } */
-        const newProduct= this.#newProduct(title,description,price,code,stock, category, status, thumbnails)
+        const newProduct= this.#newProduct(title,description,price,code,stock, category, status, thumbnails, ownerId)
         const productToInsert = new ProductDTO(newProduct)
         const errors = await this.#errorCheck(productToInsert,"add", false)
+        console.log(errors);
         return errors.length == 0 ? (await this.dao.create(productToInsert),productToInsert) : {error: errors}
         
     }
@@ -60,28 +61,33 @@ class ProductRepository{
             }
         }    
         
-    updateProductById = async (id, product) => {
+    updateProductById = async (id, product, userId) => {
         if (typeof(id) != 'object' && id.length != 24) return (console.log("ID must be 24 characters"),{error: "ID must be 24 characters"}) 
         const errors = await this.#errorCheck(product, "update", id)
-        if (!await this.dao.getOne(id)) errors.push("Product Id not found")
+        const originalProduct = await this.dao.getOne(id)
+        if (!originalProduct) errors.push("Product Id not found")
+        if ((originalProduct.owner != userId) && (userId != "admin")) errors.push("Product Owner not match")
         const updatedProduct = await this.dao.update(id,product)
         const newProduct = await this.getProductById(id)
         return errors.length == 0 ? (updatedProduct, newProduct) : errors
         
     }   
     
-    deleteProductById = async (id) => {
+    deleteProductById = async (id, userId) => {
         if (id.length == 24){
             const productToDelete = await this.dao.getOne(id)
-        if (productToDelete) return (productToDelete, await this.dao.delete(id),{message: "Success"})
-        else return {error: "Product Id not found"} 
+            if (!productToDelete) return {error: "Product Id not found"}
+            console.log(userId);
+            if ((productToDelete.owner != userId) && (userId != "admin")) return {error:"Product Owner not match"}
+            if (productToDelete) return (productToDelete, await this.dao.delete(id),{message: "Success"})
+            else return {error:"No product to delete found"} 
           } else {
               return {error:'ID must be 24 characters'}
           }
         
     }
 
-    #newProduct(title,description,price,code,stock, category, status, thumbnails){
+    #newProduct(title,description,price,code,stock, category, status, thumbnails,owner){
         const newProduct={
             title,
             description,
@@ -90,7 +96,8 @@ class ProductRepository{
             code,
             stock,
             category,
-            status
+            status,
+            owner
         }
         return newProduct;
     }
